chapter_003/src/test/java/ru/job4j/listtoarray/ConvertList2ArrayTest.java
package ru.job4j.listtoarray;

import java.util.Arrays;

import org.junit.Test;
import static org.junit.Assert.assertThat;
import static org.hamcrest.core.Is.is;

public class ConvertList2ArrayTest {

    /**
     * List<Integer> to 2D array
     */
    @Test
    public void toArray() {
        ConvertList2Array list = new ConvertList2Array();
        int[][] result = list.toArray(
                Arrays.asList(1, 2, 3, 4, 5, 6, 7), 3);
        int[][] expect = {
                {1, 2, 3},
                {4, 5, 6},
                {7, 0, 0}
        };
        assertThat(result, is(expect));
    }

}
